refactor: rename MatchControlHandler and ScoringOrchestrator classes for improved clarity and package structure
