+++ src/main/java/org/thingai/app/scoringservice/matchcontrol/Orchestrator.java
@@ -1,0 +1,4 @@
+package org.thingai.app.scoringservice.matchcontrol;
+
+public class Orchestrator {
+}
